feat: enhance comic chapter detail feature and update security configuration
- Added new endpoint in DetailComicController to fetch chapter details by slug and chapter ID.
- Updated DetailComicService to include logic for retrieving chapter details and incrementing view counts.
- Modified ChapterResponse DTO to include chapter summaries for better data structure.
- Enhanced ComicViewsHistory model to track view counts with new methods for incrementing views.
- Updated SecurityConfig to permit access to new chapter detail endpoint.
- Improved frontend components for displaying chapter details and navigation between chapters.
- Added breadcrumb navigation for better user experience in comic reading.

+++ backend-comic/src/main/java/com/thjvjpxx/backend_comic/service/impl/ChapterServiceImpl.java
@@ -81,7 +81,7 @@ public class ChapterServiceImpl implements ChapterService {
 
         List<ChapterResponse> chapterResponses = chapterList.stream()
                 .map(chapterMapper::toChapterResponse)
-                .sorted((a, b) -> Integer.compare(a.getChapterNumber(), b.getChapterNumber()))
+                .sorted((a, b) -> Double.compare(a.getChapterNumber(), b.getChapterNumber()))
                 .collect(Collectors.toList());
 
         return BaseResponse.success(
